+++ src/utils/index.js
@@ -1,3 +1,4 @@
+// @flow
 /**
  * General utilities
  * @namespace Utils
@@ -24,5 +25,7 @@ module.exports.SmartNotifierResponse = require("./smart-notifier-response");
 module.exports.PromiseReuser = require("./promise-reuser");
 
 module.exports.superagentUrlTemplatePlugin = require("./superagent-url-template-plugin");
-module.exports.superagentInstrumentationPlugin = require("././superagent-intrumentation-plugin.js");
-module.exports.superagentErrorPlugin = require("././superagent-error-plugin.js");
+module.exports.superagentInstrumentationPlugin = require("./superagent-intrumentation-plugin.js");
+module.exports.superagentErrorPlugin = require("./superagent-error-plugin.js");
+
+module.exports.devMode = require("./dev-mode");