+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.13.16
 * Explicitly handle too large kraken notification payloads
+* Adds devMode utility
+* Turn on flow in index.js files
 
 # 0.13.15
 * Bump hull-client version to 1.2.2